+++ packages/edition-slices/src/tiles/tile-e/index.js
@@ -24,6 +24,7 @@ const TileE = ({ onPress, tile, breakpoint = editionBreakpoints.small }) => {
         relativeVerticalOffset={crop.relativeVerticalOffset}
         style={styles.imageContainer}
         uri={crop.url}
+        fill
       />
       <TileSummary
         headlineStyle={styles.headline}