+++ packages/edition-slices/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.72.1](https://github.com/newsuk/times-components/compare/@times-components/edition-slices@0.72.0...@times-components/edition-slices@0.72.1) (2019-08-08)
+
+
+### Bug Fixes
+
+* REPLAT-7295 images on tiles being letterboxed ([#2165](https://github.com/newsuk/times-components/issues/2165)) ([e668cf6](https://github.com/newsuk/times-components/commit/e668cf6))
+
+
+
+
+
+# [0.72.0](https://github.com/newsuk/times-components/compare/@times-components/edition-slices@0.71.3...@times-components/edition-slices@0.72.0) (2019-08-08)
+
+
+### Features
+
+* REPLAT-7350 section item separator spacing for medium breakpoint ([#2167](https://github.com/newsuk/times-components/issues/2167)) ([6706d2f](https://github.com/newsuk/times-components/commit/6706d2f))
+
+
+
+
+
+## [0.71.3](https://github.com/newsuk/times-components/compare/@times-components/edition-slices@0.71.2...@times-components/edition-slices@0.71.3) (2019-08-07)
+
+**Note:** Version bump only for package @times-components/edition-slices
+
+
+
+
+
 ## [0.71.2](https://github.com/newsuk/times-components/compare/@times-components/edition-slices@0.71.1...@times-components/edition-slices@0.71.2) (2019-08-07)
 
 **Note:** Version bump only for package @times-components/edition-slices