+++ packages/mocks/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.36](https://github.com/newsuk/times-components/compare/@times-components/mocks@0.0.35...@times-components/mocks@0.0.36) (2019-08-07)
+
+**Note:** Version bump only for package @times-components/mocks
+
+
+
+
+
 ## [0.0.35](https://github.com/newsuk/times-components/compare/@times-components/mocks@0.0.34...@times-components/mocks@0.0.35) (2019-08-07)
 
 **Note:** Version bump only for package @times-components/mocks