+++ packages/sticky/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.8](https://github.com/newsuk/times-components/compare/@times-components/sticky@0.1.7...@times-components/sticky@0.1.8) (2019-08-07)
+
+**Note:** Version bump only for package @times-components/sticky
+
+
+
+
+
 ## [0.1.7](https://github.com/newsuk/times-components/compare/@times-components/sticky@0.1.6...@times-components/sticky@0.1.7) (2019-08-07)
 
 **Note:** Version bump only for package @times-components/sticky