+++ packages/video/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.7.66](https://github.com/newsuk/times-components/compare/@times-components/video@4.7.65...@times-components/video@4.7.66) (2019-08-07)
+
+**Note:** Version bump only for package @times-components/video
+
+
+
+
+
 ## [4.7.65](https://github.com/newsuk/times-components/compare/@times-components/video@4.7.64...@times-components/video@4.7.65) (2019-08-07)
 
 **Note:** Version bump only for package @times-components/video